refactor(admin): move the peek control out of the row behavior file
rowActions.tsx exports a hook, so adding components to it broke
react/only-export-components and cost the file its fast refresh. The
precedent cited for silencing the rule covers compound components built
with Object.assign, which genuinely cannot be split. This can.

PeekProvider, PeekTrigger and the selector move to PeekTrigger.tsx and
the file-level disable goes with them. `allowConstantExport` is on, so
the selector constant was never the violation.
